refactor(collection): update function signatures and improve ID handling

+++ internal/app/service/collection/insert.go
@@ -9,8 +9,8 @@ import (
 	"STUOJ/internal/model"
 )
 
-// 插入题单
+// Insert 插入题单
 func Insert(req request.CreateCollectionReq, reqUser model.ReqUser) (int64, error) {
-	c := collection.NewCollection(collection.WithTitle(req.Title), collection.WithDescription(req.Description), collection.WithUserId(int64(reqUser.Id)), collection.WithStatus(entity.CollectionStatus(req.Status)))
+	c := collection.NewCollection(collection.WithTitle(req.Title), collection.WithDescription(req.Description), collection.WithUserId(reqUser.Id), collection.WithStatus(entity.CollectionStatus(req.Status)))
 	return c.Create()
 }
