+++ internal/app/service/collection/misc.go
@@ -11,15 +11,15 @@ import (
 )
 
 func isPermission(c collection.Collection, reqUser model.ReqUser) error {
-	if c.UserId != int64(reqUser.Id) && reqUser.Role < entity.RoleAdmin {
+	if c.UserId != reqUser.Id && reqUser.Role < entity.RoleAdmin {
 		query := querycontext.CollectionQueryContext{}
-		query.Id.Add(int64(c.Id))
+		query.Id.Add(c.Id)
 		_, map_, err := collection.Query.SelectOne(query, collection.QueryUserId())
 		userIds, err := utils.StringToInt64Slice(map_["collection_user_id"].(string))
 		if err != nil {
 			return err
 		}
-		if !slices.Contains(userIds, int64(reqUser.Id)) {
+		if !slices.Contains(userIds, reqUser.Id) {
 			return errors.ErrUnauthorized.WithMessage("没有权限操作该题单")
 		}
 	}
